+++ src/hooks/useMounted.ts
@@ -1,0 +1,11 @@
+import { useEffect, useState } from 'react';
+
+export default function useMounted() {
+  const [mounted, setMounted] = useState<boolean>(false);
+
+  useEffect(() => {
+    setMounted(true);
+  }, []);
+
+  return mounted;
+}
